Routing code refactoring + Fix #128
